apps/api/requirements.txt
# API + Worker 共通の依存パッケージ
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Firebase Admin (認証・Firestore)
firebase-admin==6.5.0

# Google Cloud
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.0
google-cloud-pubsub==2.23.0
google-cloud-aiplatform==1.71.0

# PDF処理（Worker用）
pymupdf==1.24.0

# テスト
pytest==8.3.0
httpx==0.27.0

# その他
python-multipart==0.0.9
